置くだけプラグインの削除などを防ぐため、管理画面プラグイン一覧に置くだけプラグイン表示機能を追加

+++ src/Eccube/Service/PluginService.php
@@ -41,6 +41,36 @@ class PluginService
         $this->app = $app;
     }
 
+    /**
+     * 引数で取得した置くだけプラグインをエンティティ化
+     *
+     * @param array $developPlugins
+     * @return array $developPluginsCollection(エンティティ配列)
+     */
+    public function readDevelopPlugins(array $developPlugins){
+        // エラーハンドリング
+        if (count($developPlugins) < 1) {
+            return null;
+        }
+        // プラグイン基本チェック
+        $developPluginsCollection = array();
+        foreach ($developPlugins as $name => $path) {
+            $p = new \Eccube\Entity\Plugin();
+            $config = $this->readYml($path.'/'.self::CONFIG_YML);
+            $event = $this->readYml($path.'/'.self::EVENT_YML);
+            $p->setName(isset($config['name']) ? $config['name'] : null);
+                ->setEnable(Constant::ENABLED)
+                ->setClassName(isset($config['event']) ? $config['event'] : null)
+                ->setVersion(isset($config['version']) ? $config['version'] : null)
+                ->setDelflg(Constant::DISABLED)
+                ->setSource(0)
+                ->setCode(isset($config['code']) ? $config['code'] : null);
+
+            $developPluginsCollection[] = $p;
+        }
+        return (count($developPluginsCollection) > 0) ? $developPluginsCollection : null;
+    }
+
     public function install($path, $source = 0)
     {
         $pluginBaseDir = null;
@@ -54,6 +84,11 @@ class PluginService
 
             $config = $this->readYml($tmp.'/'.self::CONFIG_YML);
             $event = $this->readYml($tmp.'/'.self::EVENT_YML);
+
+            if (!$config && !$event) {
+                throw new PluginException('設定ファイルが見つかりませんでした');
+            }
+
             $this->deleteFile($tmp); // テンポラリのファイルを削除
 
             $this->checkSamePlugin($config['code']); // 重複していないかチェック
@@ -65,11 +100,28 @@ class PluginService
 
             $this->registerPlugin($config, $event, $source); // dbにプラグイン登録
         } catch (PluginException $e) {
-            $this->deleteDirs(array($tmp, $pluginBaseDir));
+            if (strpos('Plugin', $tmp) !== false) {
+                $removeDir = array($tmp, $pluginBaseDir);
+            }else{
+                $removeDir = array($tmp);
+            }
+            $this->deleteDirs($removeDir);
             throw $e;
         } catch (\Exception $e) { // インストーラがどんなExceptionを上げるかわからないので
-
-            $this->deleteDirs(array($tmp, $pluginBaseDir));
+            if (strpos('Plugin', $tmp) !== false) {
+                $removeDir = array($tmp, $pluginBaseDir);
+            }else{
+                $removeDir = array($tmp);
+            }
+            $this->deleteDirs($removeDir);
+            throw $e;
+        } finally {
+            if (strpos('Plugin', $tmp) !== false) {
+                $removeDir = array($tmp, $pluginBaseDir);
+            }else{
+                $removeDir = array($tmp);
+            }
+            $this->deleteDirs($removeDir);
             throw $e;
         }
